src/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2-binary
python-dotenv==0.21.1
pytest==7.4.0
pytest-django==4.5.2
django-redis==5.3.0
